Moved class RAISR to separate file raisr.py
